refactor: rename types and constants from types to wiki module
